Refactor CreateTrip components and restructure imports
- Updated import paths in createTripSteps to reflect new modular structure.
- Replaced ManualTripCreate with UpdateManualTrip in createManualTripSteps.
- Refactored AddPlaceComponent to re-export AddPlaceModal, simplifying the component structure.
- Removed obsolete DayPlanner, HorizontalDatePicker, ManualTripCreate, SpotNumber, SpotType, and WaitScreen components to streamline the codebase.
- Adjusted TripItemCard to import TypedTripItem from the new location.
