+++ features/place/presentation/components/AddPlaceModal/index.ts
@@ -1,0 +1,23 @@
+// Main component
+export { AddPlaceModal } from './AddPlaceModal'
+
+// Components
+export { ErrorState, LoadingFooter, LoadingState } from './components/LoadingStates'
+export { ModalFooter } from './components/ModalFooter'
+export { ModalHeader } from './components/ModalHeader'
+export { PlaceItem } from './components/PlaceItem'
+
+// Hooks
+export { usePlaceData } from './hooks/usePlaceData'
+export { usePlaceSelection } from './hooks/usePlaceSelection'
+
+// Types
+export type {
+  AddPlaceModalProps,
+  ErrorStateProps,
+  LoadingFooterProps,
+  LoadingStateProps,
+  ModalFooterProps,
+  ModalHeaderProps,
+  PlaceItemProps,
+} from './types'
