+++ features/trip/presentation/components/CreateTrip/AI/SpotNumber.tsx
