+++ features/trip/presentation/components/CreateTrip/AI/SpotType.tsx
