+++ features/trip/presentation/components/CreateTrip/AI/WaitScreen.tsx
